Make bool pointers always use default if nil
Previously they would first be initialized to false and then the default
applied.

If you use regular bools and set defaults after the initial population
of the struct there is no way to distinguish between an explicit false
and unset field.

This could be extended to allow the same functionality for other type
pointers, but the implementation should probably reuse existing code
better if possible.